+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Issues with parsing action strings with tuple arguments https://github.com/Textualize/textual/pull/2112
 - Fix for tabs not invalidating https://github.com/Textualize/textual/issues/2125
 - Fixed scrollbar layers issue https://github.com/Textualize/textual/issues/1358
+- Fix for interaction between pseudo-classes and widget-level render caches https://github.com/Textualize/textual/pull/2155
 
 ### Changed
 
@@ -36,6 +37,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Added auto_scroll attribute to TextLog https://github.com/Textualize/textual/pull/2127
 - Added scroll_end switch to TextLog.write https://github.com/Textualize/textual/pull/2127
+- Added `Widget.get_pseudo_class_state` https://github.com/Textualize/textual/pull/2155
 - Added Screen.ModalScreen which prevents App from handling bindings. https://github.com/Textualize/textual/pull/2139
 - Added TEXTUAL_LOG env var which should be a path that Textual will write verbose logs to (textual devtools is generally preferred) https://github.com/Textualize/textual/pull/2148
 - Added textual.logging.TextualHandler logging handler